+++ examples/common/logs_foo_library/foo_library.cc
@@ -22,7 +22,7 @@ nostd::shared_ptr<trace::Tracer> get_tracer()
 nostd::shared_ptr<logs::Logger> get_logger()
 {
   auto provider = logs::Provider::GetLoggerProvider();
-  return provider->GetLogger("foo_library_logger", "", "foo_library");
+  return provider->GetLogger("foo_library_logger", "foo_library");
 }
 }  // namespace
 
@@ -32,8 +32,8 @@ void foo_library()
   auto scoped_span = trace::Scope(get_tracer()->StartSpan("foo_library"));
   auto ctx         = span->GetContext();
   auto logger      = get_logger();
-  logger->Log(opentelemetry::logs::Severity::kDebug, "body", {}, ctx.trace_id(), ctx.span_id(),
-              ctx.trace_flags(),
-              opentelemetry::common::SystemTimestamp(std::chrono::system_clock::now()));
+  logger->EmitLogRecord(opentelemetry::logs::Severity::kDebug, "body", ctx.trace_id(),
+                        ctx.span_id(), ctx.trace_flags(),
+                        opentelemetry::common::SystemTimestamp(std::chrono::system_clock::now()));
 }
 #endif